toggle column changes, and slitted the Column definition into multiple components
